Move BiomeList to `world.biome`; move biomes v. feature to DefaultBiomes
