+++ src/main/java/amidst/gui/main/viewer/widget/CursorInformationWidget.java
@@ -13,9 +13,9 @@ import amidst.gui.main.viewer.FragmentGraphToScreenTranslator;
 import amidst.logging.AmidstLogger;
 import amidst.mojangapi.world.Dimension;
 import amidst.mojangapi.world.biome.Biome;
+import amidst.mojangapi.world.biome.BiomeList;
 import amidst.mojangapi.world.coordinates.CoordinatesInWorld;
 import amidst.mojangapi.world.coordinates.Resolution;
-import amidst.mojangapi.world.versionfeatures.BiomeList;
 import amidst.settings.Setting;
 
 @NotThreadSafe
